+++ alpacka/agents/stochastic_mcts.py
@@ -361,5 +361,5 @@ class StochasticMCTSAgent(base.OnlineAgent):
         del action_space
         return data.NetworkSignature(
             input=space_utils.space_signature(observation_space),
-            output=data.TensorSignature(shape=(1,), dtype=np.float32),
+            output=data.TensorSignature(shape=(1,)),
         )
